flask: Let UIVM run as a PVH guest
UIVM needs to be a HVM & PV to run as a PVH guest, so add the additional
xen_hvm_type.

The surfman/drm-plugin graphic stack needs to change the caching
attributes of the framebuffer region. This is done through
xc_domain_pin_memory_cacheattr which requires the dm permission
(dom0_ioemu).

This is only applied to uivm_t for now as xenfb2 is not used or
conveniently usable in a regular PVH guest.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

Squash the two commits since they are needed together to have a
functional PVH uivm.

Drop hvm_guest_map_write_grant(uivm_t) since it doesn't seem necessary
in my testing.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>